+++ packages/bpk-component-modal/src/BpkModal.js
@@ -36,14 +36,6 @@ const ScrimBpkModalDialog = withScrim(BpkModalDialog);
 type ScrimProps = {
   isIphone: boolean,
   dialogRef: () => ?HTMLElement,
-  closeEvents: {
-    onTouchStart: (event: SyntheticEvent<>) => void,
-    onTouchMove: (event: SyntheticEvent<>) => void,
-    onTouchEnd: (event: SyntheticEvent<>) => void,
-    onMouseDown: (event: SyntheticEvent<>) => void,
-    onMouseMove: (event: SyntheticEvent<>) => void,
-    onMouseUp: (event: SyntheticEvent<>) => void,
-  },
 };
 
 export type Props = {
@@ -101,11 +93,9 @@ const BpkModal = (props: Props) => {
   );
 };
 
-// isIphone, dialogRef and closeEvents are provided by the withScrim HOC
 const {
   isIphone,
   dialogRef,
-  closeEvents,
   ...modalDialogPropTypes
 } = BpkModalDialog.propTypes;